[EQL] safe test.

+++ test/test_eql/test_symbol_graph.py
@@ -8,14 +8,17 @@ from krrood.class_diagrams.utils import classes_of_module
 
 try:
     import pydot
+    import pygraphviz
 except ImportError:
     pydot = None
+    pygraphviz = None
 
 
-@pytest.mark.skipif(pydot is None, reason="pydot not installed")
+@pytest.mark.skipif(not (pydot and pygraphviz), reason="pydot not installed")
 def test_visualize_symbol_graph():
     SymbolGraph().clear()
     symbol_graph = SymbolGraph.build(classes_of_module(semantic_world_like_classes))
     symbol_graph.to_dot("symbol_graph.svg", format="svg", graph_type="type")
     assert len(symbol_graph._type_graph.wrapped_classes) == 14
-    os.remove("symbol_graph.svg")
+    if os.path.exists("symbol_graph.svg"):
+        os.remove("symbol_graph.svg")
